[feat] Reference 도메인 리네이밍(CrewReference) 및 Crews 컨텍스트 통합
이슈 번호 : #95
