+++ lib/itsm/service/request.ex
@@ -49,7 +49,7 @@ defmodule Itsm.Service.Request do
       where: [resource_type: "Request"],
       on_replace: :delete
 
-    has_many :references, Itsm.Crews.Reference,
+    has_many :references, Itsm.Crews.CrewReference,
       foreign_key: :resource_id,
       where: [resource_type: "Request"]
